# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Desafio Test Dev",
        "description": "A proposta desta aplica\u00e7\u00e3o \u00e9 efetuar o cadastro de carros de acordo com a marca, modelo e ano.",
        "termsOfService": "",
        "contact": {
            "email": "rro.oliveira@gmail.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8083",
    "basePath": "/",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
        "/api/auth/login": {
            "post": {
                "summary": "Efetua o login do usu\u00e1rio para obter o token",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "E-mail do usu\u00e1rio",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Senha do usu\u00e1rio",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exibe as informa\u00e7\u00f5es do token"
                    },
                    "401": {
                        "description": "Usu\u00e1rio sem permiss\u00e3o de acesso."
                    }
                }
            }
        },
        "/api/auth/me": {
            "get": {
                "summary": "Exibe as informa\u00e7\u00f5es do usu\u00e1rio logado.",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Bearer Token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exibe as informa\u00e7\u00f5es do usu\u00e1rio logado."
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "summary": "Efetua o logout do usu\u00e1rio.",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Bearer Token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Logout efetuado com sucesso."
                    }
                }
            }
        },
        "/api/auth/refresh": {
            "post": {
                "summary": "Atualiza o token.",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Bearer Token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna as informa\u00e7\u00f5es do token."
                    }
                }
            }
        },
        "/api/carros": {
            "get": {
                "summary": "Exibe a listagem de carros cadastrados.",
                "responses": {
                    "200": {
                        "description": "Sem resposta de retorno."
                    }
                }
            },
            "post": {
                "summary": "Efetua o cadastro do carro.",
                "parameters": [
                    {
                        "name": "marca",
                        "in": "query",
                        "description": "C\u00f3digo da marca do carro",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "modelo",
                        "in": "query",
                        "description": "Modelo do carro",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "ano",
                        "in": "query",
                        "description": "Ano de fabrica\u00e7\u00e3o do carro",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Carro cadastrado com sucesso."
                    },
                    "400": {
                        "description": "\n     *                                              O c\u00f3digo da marca informado est\u00e1 incorreto.\n     *                                              O ano do carro informado est\u00e1 incorreto.\n     *                                              Erro ao efetuar o cadastro do carro.\n     *                                              "
                    }
                }
            }
        },
        "/api/carros/{id}": {
            "get": {
                "summary": "Exibe as informa\u00e7\u00f5es do carro de acordo com o id informado",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "C\u00f3digo do register do carro",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sem resposta de retorno."
                    },
                    "400": {
                        "description": "\n     *                                              Erro ao localizar o register do carro informado.\n     *                                              "
                    }
                }
            },
            "put": {
                "summary": "Efetua atualiza\u00e7\u00e3o das informa\u00e7\u00f5es do carro.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "C\u00f3digo do register do carro",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "marca",
                        "in": "query",
                        "description": "C\u00f3digo da marca do carro",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "modelo",
                        "in": "query",
                        "description": "Modelo do carro",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "ano",
                        "in": "query",
                        "description": "Ano de fabrica\u00e7\u00e3o do carro",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Carro atualizado com sucesso."
                    },
                    "400": {
                        "description": "\n     *                                              O c\u00f3digo da marca informado est\u00e1 incorreto.\n     *                                              O ano do carro informado est\u00e1 incorreto.\n     *                                              Erro ao localizar o register do carro informado.\n     *                                              Erro ao atualizar as informa\u00e7\u00f5es do carro.\n     *                                              "
                    }
                }
            },
            "delete": {
                "summary": "Efetua a exclus\u00e3o do cadastro do carro.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "C\u00f3digo do register do carro",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Carro exclu\u00eddo com sucesso."
                    },
                    "400": {
                        "description": "\n     *                                              Erro ao localizar o register do carro informado.\n     *                                              "
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "summary": "Lista todos os usu\u00e1rios da aplica\u00e7\u00e3o.",
                "responses": {
                    "200": {
                        "description": "Listagem com todos os usu\u00e1rios da aplica\u00e7\u00e3o."
                    }
                }
            }
        }
    },
    "definitions": {}
}